ml_project/requirements.txt
matplotlib==3.4.1
tabulate==0.8.9
hydra-core==1.0.6
marshmallow_dataclass==8.4.1
numpy==1.20.2
pandas==1.2.4
PyYAML==5.4.1
scikit_learn==0.24.2
typeguard==2.12.0
catboost==1.0.5